Organize activities and fragments in different packages
