+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
             android:roundIcon="@mipmap/ic_launcher_round"
             android:supportsRtl="true"
             android:theme="@style/AppTheme">
-        <activity android:name=".GamesListActivity"
+        <activity android:name=".activities.GamesListActivity"
                   android:label="@string/app_name"
                   android:screenOrientation="portrait"
                   android:theme="@style/AppTheme.NoActionBar">
@@ -22,7 +22,7 @@
             </intent-filter>
         </activity>
 
-        <activity android:name=".GameDisplayActivity" android:parentActivityName=".GamesListActivity" android:screenOrientation="portrait" />
+        <activity android:name=".activities.GameDisplayActivity" android:parentActivityName=".activities.GamesListActivity" android:screenOrientation="portrait" />
 
     </application>
