+++ app/src/main/res/layout/activity_game_display.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context="rocks.massi.trollsgames.GameDisplayActivity">
+        tools:context="rocks.massi.trollsgames.activities.GameDisplayActivity">
 
 
     <ImageView
