+++ app/src/main/res/layout/app_bar_games_list.xml
@@ -5,7 +5,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context="rocks.massi.trollsgames.GamesListActivity">
+        tools:context="rocks.massi.trollsgames.activities.GamesListActivity">
 
     <android.support.design.widget.AppBarLayout
             android:layout_height="wrap_content"
